Add reusable theme, styleguide, and project for debug drawing utilities
The goal is to compile reusable components and utilities to draw in the viewport, ingame, but also compose debug interfaces fast and efficiently.
I'm putting it in the repo for now as it's small but later we can move it to another repo and have it as a git submodule, or turn it in an official addon.